dev <- main

+++ secureworkload/resource_secureworkload_scope.go
@@ -2,12 +2,13 @@ package secureworkload
 
 import (
 	"fmt"
-
+	"time"
+	"strings"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	// client "github.com/secureworkload-exchange/terraform-go-sdk"
 	// secureworkload "github.com/secureworkload-exchange/terraform-go-sdk"
 )
-
+var timer int
 func resourceSecureWorkloadScope() *schema.Resource {
 	return &schema.Resource{
 		Description: "Resource for creating a scope in Secure Workload\n" +
@@ -209,5 +210,18 @@ func resourceSecureWorkloadScopeUpdate(d *schema.ResourceData, meta interface{})
 
 func resourceSecureWorkloadScopeDelete(d *schema.ResourceData, meta interface{}) error {
 	client := meta.(Client)
-	return client.DeleteScope(d.Id())
+	err :=  client.DeleteScope(d.Id())
+	for(err != nil){
+		if(strings.Contains(err.Error(), "error:cannot delete scope because it is in use")){
+			if(timer >= 20 ){
+				return err
+			}
+			time.Sleep(60 * time.Second)
+			timer = timer+1
+			err = client.DeleteScope(d.Id())
+		}else {
+			return err
+		}
+	}
+	return err
 }
